Fix bug with negative times
The cause was that we were relying on the date of the most recent action in the case where there were no matching entries/exits to/from lists. The fix is to rely on the oldest action date instead, which would make it impossible for the computed times to be negative. Note that this does not mean that the times are accurate, since changing list suffixes will still cause problems